requirement.txt
flask
flask-cors
gunicorn
psycopg2-binary
python-dotenv
openai
flask-restx
PyYAML